// src/api/admin/AdminType.ts
export type AdminMode = {
    aid: string;
    roid: [];
    account: string;
    apassword: string;
    aemail: string;
    aportrait: string;
    admin: string;
    asex:string;
    aphone:string;
    aname: string;
    type:string;
}

export type AdminParm = {
    pageSize: number;
    currentPage: number;
    account: string;
}